feat: add main service

+++ main-service/main-server/src/main/java/ru/practicum/event/controller/AdminEventServerController.java
@@ -22,9 +22,9 @@ public class AdminEventServerController {
 
     @GetMapping
     public List<EventResponseDto>  searchEvents(
-            @RequestParam(required = false) List<Integer> users,
+            @RequestParam(required = false) List<Long> users,
             @RequestParam(required = false) List<EventStatus> states,
-            @RequestParam(required = false) List<Integer> categories,
+            @RequestParam(required = false) List<Long> categories,
             @RequestParam(required = false) String rangeStart,
             @RequestParam(required = false) String rangeEnd,
             @RequestParam(defaultValue = "0") Integer from,
